requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-groq
langgraph
docker
python-dotenv
pytest
PyGithub
gitpython
duckduckgo-search
flake8
networkx
fastapi
uvicorn[standard]
ddgs
langchain-openai